SPARK-40034. PathOutputCommitters to support dynamic partition overwrite.
Uses the StreamCapabilities probe in MAPREDUCE-7403 to identify when a
PathOutputCommitter is compatible with dynamic partition overwrite.

This patch has unit tests but not integration tests; really needs
to test the SQL commands through the manifest committer into gcs/abfs,
or at least local fs. That would be possible once hadoop 3.3.5 is out...
Change-Id: I5cbc391bc021b4dd177374e82de9fc33137ac319